Prefab Map, Enemy AI, Modular FPS Controller
Prefab map imported and tweaked, enemy AI has been created. Using the AI agent and AI Mesh navigator the zombies plot a path to the player's position. The old FPS system was changed to the Modular FPS Controller plugin which honed in on the smoothness of the controls as well as adding crouch and zoom.